chore: clean up + modularize code
